refactor: set name according google style
