+++ src/preprocessor/boundaries/force.cpp
@@ -1,7 +1,7 @@
 #include <optional>
 #include <unordered_map>
-#include "../error.h"
-#include "Force.h"
+#include "error.h"
+#include "force.h"
 
 Force::Force(int i, float x, float y) : index(i), force_x(x), force_y(y) {};
