+++ src/preprocessor/boundaries/support.cpp
@@ -1,6 +1,6 @@
 #include <unordered_map>
-#include "Support.h"
-#include "../error.h"
+#include "support.h"
+#include "error.h"
 #include <optional>
 
 Support::Support(int index, std::optional<double> disp_x, std::optional<double> disp_y):
