+++ src/preprocessor/preprocessor.cpp
@@ -6,8 +6,8 @@
 #include "elem_service/ielement.h"
 #include "node/node.h"
 #include "material/material.h"
-#include "boundaries/Force.h"
-#include "boundaries/Support.h"
+#include "boundaries/force.h"
+#include "boundaries/support.h"
 #include "elem_service/elem_params.h"
 #include "elem_service/elem_creator.h"
 #include "tools.h"
